+++ packages/global-styles-engine/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### New Features
 
 -   Add text shadow settings and presets ([#79584](https://github.com/WordPress/gutenberg/pull/79584)).
+-   Accept `background.backgroundClip` as a valid theme.json setting ([#77141](https://github.com/WordPress/gutenberg/pull/77141)).
 
 ### Bug Fixes